.history/requirements_20250518113702.txt
pandas
numpy
matplotlib
scikit-learn
pytz
pytest
python-dotenv